project/using-flask-restx/requirements.txt
Flask-JWT-Extended
Flask-RESTX
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv